Update README.md
The improved version has a loop where players can keep playing the game for as long as they want to. The game also indicates the scores for both players and, if either players reach the score of 3 before the other, the game restarts. The players also have the option of exiting the game completely if they want to.